back/requirements.txt
# Django and REST framework
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-filter==23.3

# Database
psycopg[binary]==3.2.6  # Updated to latest compatible version
dj-database-url==2.1.0

# Image processing
Pillow==11.2.0  # Updated to version compatible with Python 3.13

# Environment variables
python-dotenv==1.0.0

# Development tools
black==23.10.1
flake8==6.1.0

# Production
gunicorn==21.2.0
whitenoise==6.5.0